Feedback iteration (pass 1)
  * remove plugins, ignore directory

  * shader to spaces

  * refactor IPeerImpl => IPeerInteraction

  * namespace => WebRTC

  * nit: android spacing fix

  * nit: WebrtcPeerVideo => spaces

  * add managed i420a walk comments

  * remove ui controls (unclear sample)

  * fix typo 'genetic'

  * document android dllPath change

  * ISignaler: add connect/disconnect event

  * licensing pass (add ms copyright)
