+++ Assets/MixedRealityToolkit.Extensions/Webrtc/.gitignore
@@ -1,0 +1,2 @@
+# Ignore Plugins, as those assets cannot be checked in
+Plugins/
