--- Assets/MixedRealityToolkit.Extensions/Webrtc/Plugins/.gitignore
@@ -1,1 +1,0 @@
-*.dll
