--- Assets/MixedRealityToolkit.Extensions/Webrtc/Plugins/WSA.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bda057fb83af1dd4f9c9b5c78a1820e0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
