--- Assets/MixedRealityToolkit.Extensions/Webrtc/Plugins/WSA/x86.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b258b3d4021354943ae1ad544ff89557
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
