--- Assets/MixedRealityToolkit.Extensions/Webrtc/Plugins/WSA/x86/webrtc_unity_plugin.dll.meta
@@ -1,32 +1,0 @@
-fileFormatVersion: 2
-guid: 38c788abea344ea45a7be4f17d263a1b
-PluginImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  iconMap: {}
-  executionOrder: {}
-  defineConstraints: []
-  isPreloaded: 0
-  isOverridable: 0
-  isExplicitlyReferenced: 0
-  platformData:
-  - first:
-      Any: 
-    second:
-      enabled: 0
-      settings: {}
-  - first:
-      Editor: Editor
-    second:
-      enabled: 0
-      settings:
-        DefaultValueInitialized: true
-  - first:
-      Windows Store Apps: WindowsStoreApps
-    second:
-      enabled: 1
-      settings:
-        CPU: x86
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
