--- Assets/MixedRealityToolkit.Extensions/Webrtc/Plugins/WSA/x86_64.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3d134a611cd99ce408b53854f0db518c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
