--- Assets/MixedRealityToolkit.Extensions/Webrtc/Plugins/Win32.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 78cdce230068e264b9358d9661161eeb
-folderAsset: yes
-timeCreated: 1535763114
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
