--- Assets/MixedRealityToolkit.Extensions/Webrtc/Plugins/Win32/x64.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 3072722b35e18cf4c8c72d2509eaca30
-folderAsset: yes
-timeCreated: 1535763124
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
