--- Assets/MixedRealityToolkit.Extensions/Webrtc/Plugins/Win32/x86.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: b9aef6e9884be5c46a9cc50a934a043b
-folderAsset: yes
-timeCreated: 1535763129
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
