+++ Assets/MixedRealityToolkit.Extensions/Webrtc/Scripts/Delegates/I420FrameReadyHandler.cs
@@ -1,6 +1,8 @@
-﻿using System;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+using System;
 
-namespace Microsoft.MixedReality.Toolkit.Extensions.Webrtc.Delegates
+namespace Microsoft.MixedReality.Toolkit.Extensions.WebRTC.Delegates
 {
     /// <summary>
     /// I420 frame callback
