+++ Assets/MixedRealityToolkit.Extensions/Webrtc/Scripts/Editor/WebRtcVideoPlayerEditor.cs
@@ -1,7 +1,9 @@
-﻿using UnityEditor;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+using UnityEditor;
 using UnityEngine;
 
-namespace Microsoft.MixedReality.Toolkit.Extensions.Webrtc.Editor
+namespace Microsoft.MixedReality.Toolkit.Extensions.WebRTC.Editor
 {
 #if UNITY_EDITOR
     /// <summary>
