+++ Assets/MixedRealityToolkit.Extensions/Webrtc/Scripts/IPeerInteraction.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2569aee58ef58a84ba96698c1bdadb09
+guid: cb3c5bb2c6d058b49a05145440b487cb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
