+++ Assets/MixedRealityToolkit.Extensions/Webrtc/Scripts/Marshalling/Deque.cs
@@ -3,7 +3,7 @@
 namespace System.Collections.Generic
 {
     /// <summary>
-    /// A genetic Deque class. It can be thought of as
+    /// A generic Deque class. It can be thought of as
     /// a double-ended queue, hence Deque. This allows for
     /// an O(1) AddFront, AddBack, RemoveFront, RemoveBack.
     /// The Deque also has O(1) indexed lookup, as it is backed
