+++ Assets/MixedRealityToolkit.Extensions/Webrtc/Scripts/Marshalling/IReadonlyFrameQueueStat.cs
@@ -1,4 +1,6 @@
-﻿namespace Microsoft.MixedReality.Toolkit.Extensions.Webrtc.Marshalling
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+namespace Microsoft.MixedReality.Toolkit.Extensions.WebRTC.Marshalling
 {
     /// <summary>
     /// A readonly frame queue stat
