+++ Assets/MixedRealityToolkit.Extensions/Webrtc/Scripts/Signaling/SignalerMessage.cs
@@ -1,7 +1,9 @@
-﻿using System;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+using System;
 using UnityEngine;
 
-namespace Microsoft.MixedReality.Toolkit.Extensions.Webrtc.Signaling
+namespace Microsoft.MixedReality.Toolkit.Extensions.WebRTC.Signaling
 {
     /// <summary>
     /// Data that makes up a signaler message
